pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "matplotlib>=3.10.1",
    "numpy==1.24.3",
    "pillow>=11.1.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "tensorflow>=2.14.0",
    "twilio>=9.5.1",
]
